+++ sotodlib/mapmaking/__init__.py
@@ -1,7 +1,7 @@
-from .utils import utils
-from .mapmakers import ml_mapmaker, demod_mapmaker
-from .noise_model import *
-from .pointing_matrix import *
-from .obs_grouping import *
 from .bench import *
 from .log import *
+from .mapmakers import demod_mapmaker, ml_mapmaker
+from .noise_model import *
+from .obs_grouping import *
+from .pointing_matrix import *
+from .utils import utils
