+++ sotodlib/mapmaking/signals/__init__.py
@@ -1,0 +1,2 @@
+from .demod_signals import (DemodSignal, DemodSignalMap,  # noqa: F401
+                            add_weights_to_info)
