+++ sotodlib/mapmaking/utils/log.py
@@ -6,15 +6,20 @@
 
 """This module provids a simple logging/output interface."""
 import time
-from logging import *
-from pixell import memory, colors, mpi
+from logging import (DEBUG, ERROR, INFO, FileHandler, Filter, Formatter,
+                     StreamHandler, getLogger)
+
+from pixell import colors, memory, mpi
+
 
 class EnFilter(Filter):
     def __init__(self, rank=0):
         self.rank = rank
         try:
             # Try to get actual time since task start if possible
-            import os, psutil
+            import os
+
+            import psutil
             p = psutil.Process(os.getpid())
             self.t0 = p.create_time()
         except ImportError:
@@ -62,7 +67,7 @@ def init(level=INFO, rank=mpi.COMM_WORLD.rank, file=None, fmt=default_format, co
     The console level threshold is set by the level argument, which must be a
     python logging module level. The threshold does not apply to file output, where
     everything is output."""
-    logger  = getLogger("enlib")
+    logger  = getLogger("sotodlib")
     logger.setLevel(DEBUG)
     if file:
         try:
