+++ sotodlib/mapmaking/utils/utils.py
@@ -1,12 +1,3 @@
-from typing import Optional, Any, Union
-from sqlalchemy import create_engine, exc
-from sqlalchemy.orm import declarative_base, Mapped, mapped_column, sessionmaker
-import importlib
-import numpy as np
-import so3g
-from pixell import enmap, fft, resample, tilemap, bunch, utils as putils
-
-from ... import coords, core, tod_ops
 
 
 def deslope_el(tod, el, srate, inplace=False):
