+++ sotodlib/site_pipeline/make_ml_map.py
@@ -34,13 +34,11 @@ def get_parser(parser=None):
 
 def main(**args):
     import numpy as np, sys, time, warnings, os, so3g
-    from sotodlib.core import Context, AxisManager, IndexAxis
-    from sotodlib.io import metadata   # PerDetectorHdf5 work-around
+    from sotodlib.core import Context
     from sotodlib import tod_ops, mapmaking, core
     from sotodlib.tod_ops import filters
     from sotodlib.mapmaking import log
     from pixell import enmap, utils, fft, bunch, wcsutils, mpi, bench
-    import yaml
 
     #try: import moby2.analysis.socompat
     #except ImportError: warnings.warn("Can't import moby2.analysis.socompat. ACT data input probably won't work")
